+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.3] - 2026-08-29
  
 ### Added
+- Hotfix: Resolve base branch to the merge destination instead of current branch
 - Add validation for args options in pytest.ini
 - Add default values for diff-level, rank-level and no-rank flags in pytest.ini
 ## [0.6.2] - 2026-08-22